feat(standard_layout): Refactor feedfollows to internal/
