--- handler_feed_follows.go
@@ -1,112 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/go-chi/chi"
-	"github.com/google/uuid"
-	"github.com/mike-kimani/fechronizo/internal/database"
-)
-
-func (apiCfg *apiConfig) handlerCreateFeedFollow(w http.ResponseWriter, r *http.Request, user database.User) {
-
-	type parameters struct {
-		FeedID uuid.UUID `json:"feed_id"`
-	}
-
-	decoder := json.NewDecoder(r.Body)
-	params := parameters{}
-
-	err := decoder.Decode(&params)
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("error parsing json %v", err))
-		return
-	}
-
-	feedFollow, err := apiCfg.DB.CreateFeedFollows(r.Context(), database.CreateFeedFollowsParams{
-		ID:        uuid.New(),
-		CreatedAt: time.Now(),
-		UpdatedAt: time.Now(),
-		UserID:    user.ID,
-		FeedID:    params.FeedID,
-	})
-
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("Couldn't create feed follow %v", err))
-	}
-
-	respondWithJson(w, 200, databaseFeedFollowtoFeedFollow(feedFollow))
-
-}
-
-func (apiCfg *apiConfig) handlerGetFeedFollows(w http.ResponseWriter, r *http.Request, user database.User) {
-	type parameters struct {
-		UserID uuid.UUID `json:"user_id"`
-	}
-
-	decoder := json.NewDecoder(r.Body)
-	params := parameters{}
-
-	err := decoder.Decode(&params)
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("error parsing json %v", err))
-		return
-	}
-
-	feedFollows, err := apiCfg.DB.GetFeedFollows(r.Context(), params.UserID)
-
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("Couldn't get feed follows %v", err))
-		return
-	}
-
-	respondWithJson(w, 200, databaseFeedFollowstoFeedFollows(feedFollows))
-}
-
-func (apiCfg *apiConfig) handlerDeleteFeedFollows(w http.ResponseWriter, r *http.Request, user database.User) {
-	type parameters struct {
-		FeedID uuid.UUID `json:"feed_follow_id"`
-	}
-
-	decoder := json.NewDecoder(r.Body)
-	params := parameters{}
-
-	err := decoder.Decode(&params)
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("error parsing json %v", err))
-		return
-	}
-
-	err = apiCfg.DB.DeleteFeedFollows(r.Context(), database.DeleteFeedFollowsParams{
-		ID:     params.FeedID,
-		UserID: user.ID,
-	})
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("couldn't delete feed follow %v", err))
-		return
-	}
-	respondWithJson(w, 200, "Successfully deleted feed follow")
-}
-
-func (apiCfg *apiConfig) handlerDeleteFeedFollows2(w http.ResponseWriter, r *http.Request, user database.User) {
-	feedFollowIDStr := chi.URLParam(r, "feedFollowID")
-	feedFollowID, err := uuid.Parse(feedFollowIDStr)
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("couldn't parse the feed ID %v", err))
-		return
-	}
-	err = apiCfg.DB.DeleteFeedFollows(r.Context(), database.DeleteFeedFollowsParams{
-		ID:     feedFollowID,
-		UserID: user.ID,
-	})
-
-	if err != nil {
-		respondWithError(w, 400, fmt.Sprintf("couldn't delete feed follow %v", err))
-		return
-	}
-
-	respondWithJson(w, 200, struct{}{})
-}
